Factor out entities and inits.
